Create leapYear.java
Given the year, write a program that outputs 1 or 0 for leap years.

A leap year is when the year is a multiple of 4, not a multiple of 100, or a multiple of 400.

For example, 2012 is a leap year because it is a multiple of 4 and not a multiple of 100. 1900 is not a leap year because it is a multiple of 100 and not a multiple of 400. However, 2000 is a leap year because it is a multiple of 400.