Create 1110.java
When an integer greater than or equal to 0 and less than or equal to 99 is given, the following operation can be performed. First, if the given number is less than 10, add zero to the front and add the numbers of each digit. Then, we can add the rightmost digit of a given number and the rightmost digit of the sum previously obtained to create a new number. Let's look at the next example.

It starts at 26. 2+6 = 8. The new number is 68. 6+8 = 14. The new number is 84. 8+4 = 12. The new number is 42. 4+2 = 6. The new number is 26.

The above example can be returned to the original number in four times. Thus, the length of the cycle 26 is 4.

Given N, write a program to find the length of the cycle of N.