Create 1712.java
World Electronics is a company that manufactures and sells laptops. Regardless of the number of laptops sold, it costs a fixed cost of A million won every year, including rent, property tax, insurance premiums, and salaries, and it costs a total of B million won, including material and labor costs.

For example, let's say A = 1,000 and B = 70. In this case, it costs a total of 10.7 million won to produce a single laptop, and a total of 17 million won to produce tropical products.

It is said that the price of the laptop was set at C million won. In general, as the number of productions increases, at some point, the total income (sales cost) will be higher than the total cost (= fixed cost + variable cost). For the first time, the point where profits are generated because gross income exceeds total cost is called the BREAK-EVEN POINT.

Given A, B, and C, write a program to find a break-even point.