Create 15552.java
There is something to be careful about before solving the for question in earnest. If the input/output method is slow, time may exceed when receiving or outputting multiple lines.

If you are using C++ and want to use c/cout, apply both c.tie (NULL) and sync_with_stdio (false), and write an opening character (\n) instead of endl. However, in this case, the input/output method of C, such as scanf/printf/puts/getchar/putchar, should no longer be used.

If you are using Java, you can use BufferedReader and BufferedWriter instead of Scanner and System.out.println. BufferedWriter.flush only needs to be done once at the end.

If Python is being used, sys.stdin.readline can be used instead of input. However, in this case, even the opening character at the end is input together, so if you want to save the character string, it is recommended to add .rstrip().

In addition, since the input and output streams are separate, it is not necessary to receive, save, and output all test cases. You can print one out after receiving one test case.

The number T of test cases is given in the first row. T is up to 1,000,000. The next line T is given two integers A and B, respectively. A and B are 1 or more and 1,000 or less.