Create 11021.java
Enter the two integers A and B, and then write a program that outputs A+B.
The number of test cases T is given in the first row.

Each test case consists of a single line, and A and B are given to each line. The input consists of several test cases. Each test case outputs "Case #x:", and then outputs A+B. The test case number starts from 1.