Create 11022.java
Enter the two integers A and B, and then write a program that outputs A+B.
The number of test cases T is given in the first row.
Each test case consists of a single line, and A and B are given to each line. (0 < A, B < 10)
Each test case is output in the form of "Case #x: A + B = C". x is the test case number and starts with 1, and C is A+B.